step04/Feign: Add Feign support. - Create a « sales-service » MS ; - Retrieve sales in « client-service » which uses a Feign client (SalesClient) to call « sales-service » which exposes the path « /api/sales/client/{clientId} ».
