+++ client-service/src/main/java/ma/hero/clients/ClientServiceApplication.java
@@ -1,0 +1,17 @@
+package ma.hero.clients;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@SpringBootApplication
+@EnableSwagger2
+@EnableFeignClients
+public class ClientServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ClientServiceApplication.class, args);
+	}
+
+}
