+++ client-service/src/main/java/ma/hero/clients/repository/ClientRepository.java
@@ -1,0 +1,13 @@
+package ma.hero.clients.repository;
+
+import ma.hero.clients.domain.Client;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+/**
+ * Spring Data repository for the Client entity.
+ */
+@Repository
+public interface ClientRepository extends JpaRepository<Client, Long> {
+}
