--- client-service/src/main/java/ma/s2m/clients/ClientServiceApplication.java
@@ -1,15 +1,0 @@
-package ma.s2m.clients;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@SpringBootApplication
-@EnableSwagger2
-public class ClientServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ClientServiceApplication.class, args);
-	}
-
-}
