--- client-service/src/main/java/ma/s2m/clients/config/KeycloakSecurityConfiguration.java
@@ -1,96 +1,0 @@
-package ma.s2m.clients.config;
-
-import org.keycloak.adapters.KeycloakConfigResolver;
-import org.keycloak.adapters.springboot.KeycloakBaseSpringBootConfiguration;
-import org.keycloak.adapters.springboot.KeycloakSpringBootConfigResolver;
-import org.keycloak.adapters.springsecurity.KeycloakConfiguration;
-import org.keycloak.adapters.springsecurity.authentication.KeycloakAuthenticationProvider;
-import org.keycloak.adapters.springsecurity.config.KeycloakWebSecurityConfigurerAdapter;
-import org.keycloak.adapters.springsecurity.filter.KeycloakAuthenticationProcessingFilter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.core.authority.mapping.SimpleAuthorityMapper;
-import org.springframework.security.web.authentication.session.NullAuthenticatedSessionStrategy;
-import org.springframework.security.web.authentication.session.SessionAuthenticationStrategy;
-
-@KeycloakConfiguration
-@EnableGlobalMethodSecurity(prePostEnabled = true, proxyTargetClass = true)
-public class KeycloakSecurityConfiguration extends KeycloakWebSecurityConfigurerAdapter {
-    @Configuration
-    static class CustomKeycloakBaseSpringBootConfiguration extends KeycloakBaseSpringBootConfiguration {
-    }
-
-    /**
-     * Defines the session authentication strategy.
-     */
-    @Bean
-    @Override
-    protected SessionAuthenticationStrategy sessionAuthenticationStrategy() {
-        // required for bearer-only applications.
-        return new NullAuthenticatedSessionStrategy();
-    }
-
-    /**
-     * Prevent double bean declaration
-     */
-    @SuppressWarnings({ "rawtypes", "unchecked" })
-    @Bean
-    public FilterRegistrationBean keycloakAuthenticationProcessingFilterRegistrationBean(KeycloakAuthenticationProcessingFilter filter) {
-        FilterRegistrationBean registrationBean = new FilterRegistrationBean(filter);
-        registrationBean.setEnabled(false);
-        return registrationBean;
-    }
-
-    /**
-     * Registers the KeycloakAuthenticationProvider with the authentication manager.
-     */
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) {
-        KeycloakAuthenticationProvider keycloakAuthenticationProvider = new KeycloakAuthenticationProvider();
-        // simple Authority Mapper to avoid ROLE_
-        keycloakAuthenticationProvider.setGrantedAuthoritiesMapper(new SimpleAuthorityMapper());
-        auth.authenticationProvider(keycloakAuthenticationProvider);
-    }
-
-    /**
-     * Required to handle spring boot configurations
-     *
-     * @return
-     */
-    @Bean
-    public KeycloakConfigResolver keycloakConfigResolver() {
-        return new KeycloakSpringBootConfigResolver();
-    }
-
-    /**
-     * Keycloak specific configuration
-     *
-     * @param http
-     * @throws Exception
-     */
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        super.configure(http);
-        // add cors options
-        http.cors().and().csrf().disable();
-        http.authorizeRequests().antMatchers("/api/**").authenticated();
-
-        // No security restriction for static resources : swagger ui & spring fox
-        configureStaticResourcesAccessControlRules(http);
-
-    }
-
-    private void configureStaticResourcesAccessControlRules(HttpSecurity http) throws Exception {
-        http.authorizeRequests().antMatchers("/v2/api-docs/**").permitAll();
-        http.authorizeRequests().antMatchers("/swagger-resources/**").permitAll();
-        http.authorizeRequests().antMatchers("/swagger-ui.html").permitAll();
-        http.authorizeRequests().antMatchers("/webjars/**/*").permitAll();
-        http.authorizeRequests().antMatchers("/swagger-resources/webjars/**").permitAll();
-    }
-
-}
