+++ client-service/src/test/java/ma/hero/clients/ClientServiceApplicationTests.java
@@ -1,0 +1,14 @@
+package ma.hero.clients;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+
+@SpringBootTest
+class ClientServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
