+++ client-service/src/test/java/ma/hero/clients/config/WithMockOAuth2Context.java
@@ -1,0 +1,13 @@
+package ma.hero.clients.config;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+import org.springframework.security.test.context.support.WithSecurityContext;
+
+@Retention(RetentionPolicy.RUNTIME)
+@WithSecurityContext(factory = WithMockOAuth2SecurityContextFactory.class)
+public @interface WithMockOAuth2Context {
+
+    String authorities() default "";
+}
