--- client-service/src/test/java/ma/s2m/clients/ClientServiceApplicationTests.java
@@ -1,14 +1,0 @@
-package ma.s2m.clients;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-
-@SpringBootTest
-class ClientServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
