--- client-service/src/test/java/ma/s2m/clients/config/WithMockOAuth2Context.java
@@ -1,13 +1,0 @@
-package ma.s2m.clients.config;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-import org.springframework.security.test.context.support.WithSecurityContext;
-
-@Retention(RetentionPolicy.RUNTIME)
-@WithSecurityContext(factory = WithMockOAuth2SecurityContextFactory.class)
-public @interface WithMockOAuth2Context {
-
-    String authorities() default "";
-}
