+++ sales-service/src/test/java/ma/hero/sales/SaleServiceApplicationTests.java
@@ -1,0 +1,16 @@
+package ma.hero.sales;
+
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+class SaleServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
